Fix compiler warnings

+++ src/main/WorkScheduleEditorPresenterProxy.h
@@ -62,8 +62,8 @@ private:
     void onAddExceptionalRequested() override;
 
     void onExceptionalDaysAdded(dw::Date startDate,
-                                uint16_t numDays,
-                                uint16_t sprintsPerDay) override;
+                                int32_t numDays,
+                                int32_t sprintsPerDay) override;
 
     void onExceptionalDayRemoved(dw::Date date) override;
 
@@ -125,7 +125,7 @@ inline void WorkScheduleEditorPresenterProxy::onAddExceptionalRequested()
 }
 
 inline void WorkScheduleEditorPresenterProxy::onExceptionalDaysAdded(
-    dw::Date startDate, uint16_t numDays, uint16_t sprintsPerDay)
+    dw::Date startDate, int32_t numDays, int32_t sprintsPerDay)
 {
     presenter.onExceptionalDaysAdded(startDate, numDays, sprintsPerDay);
 }
